+++ test/unit/project-management/project-management.spec.ts
@@ -22,7 +22,8 @@ import * as sinon from 'sinon';
 import { FirebaseApp } from '../../../src/firebase-app';
 import { AndroidApp } from '../../../src/project-management/android-app';
 import { ProjectManagement } from '../../../src/project-management/project-management';
-import { ProjectManagementRequestHandler } from '../../../src/project-management/project-management-api-request';
+import { ProjectManagementImpl } from '../../../src/project-management/project-management-internal';
+import { ProjectManagementRequestHandler } from '../../../src/project-management/project-management-api-request-internal';
 import { FirebaseProjectManagementError } from '../../../src/utils/error';
 import * as mocks from '../../resources/mocks';
 import { IosApp } from '../../../src/project-management/ios-app';
@@ -59,7 +60,7 @@ describe('ProjectManagement', () => {
   beforeEach(() => {
     mockApp = mocks.app();
     mockCredentialApp = mocks.mockCredentialApp();
-    projectManagement = new ProjectManagement(mockApp);
+    projectManagement = new ProjectManagementImpl(mockApp);
   });
 
   afterEach(() => {
@@ -73,7 +74,7 @@ describe('ProjectManagement', () => {
     invalidApps.forEach((invalidApp) => {
       it('should throw given invalid app: ' + JSON.stringify(invalidApp), () => {
         expect(() => {
-          const projectManagementAny: any = ProjectManagement;
+          const projectManagementAny: any = ProjectManagementImpl;
           return new projectManagementAny(invalidApp);
         }).to.throw(
           'First argument passed to admin.projectManagement() must be a valid Firebase app '
@@ -83,7 +84,7 @@ describe('ProjectManagement', () => {
 
     it('should throw given no app', () => {
       expect(() => {
-        const projectManagementAny: any = ProjectManagement;
+        const projectManagementAny: any = ProjectManagementImpl;
         return new projectManagementAny();
       }).to.throw(
         'First argument passed to admin.projectManagement() must be a valid Firebase app '
@@ -94,14 +95,14 @@ describe('ProjectManagement', () => {
       // Project ID not set in the environment.
       delete process.env.GOOGLE_CLOUD_PROJECT;
       delete process.env.GCLOUD_PROJECT;
-      const projectMgt = new ProjectManagement(mockCredentialApp);
+      const projectMgt = new ProjectManagementImpl(mockCredentialApp);
       return projectMgt.listIosApps()
         .should.eventually.rejectedWith(noProjectIdErrorMessage);
     });
 
     it('should not throw given a valid app', () => {
       expect(() => {
-        return new ProjectManagement(mockApp);
+        return new ProjectManagementImpl(mockApp);
       }).not.to.throw();
     });
   });
